moved some parsing methods to other file
