+++ lib/utils/parsing_utils.dart
@@ -1,7 +1,12 @@
 import 'dart:convert';
+import 'dart:developer';
+import 'dart:typed_data';
 
 import 'package:asn1lib/asn1lib.dart';
 import 'package:pointycastle/export.dart';
+import 'package:privacyidea_authenticator/utils/utils.dart';
+
+import 'identifiers.dart';
 
 /// Extract RSA-Public-Keys from DER structure that is a BASE64 encoded Strings.
 /// According to the PKCS#1 format:
@@ -162,4 +167,284 @@ RSAPrivateKey deserializeRSAPrivateKeyPKCS1(String keyStr) {
   BigInt q = (asn1sequence.elements[5] as ASN1Integer).valueAsBigInteger;
 
   return RSAPrivateKey(modulus, exponent, p, q);
+}
+
+/// The method returns a map that contains all the uri parameters.
+Map<String, dynamic> parseQRCodeToMap(String uriAsString) {
+  Uri uri = Uri.parse(uriAsString);
+  log(
+    "Barcode is valid Uri:",
+    name: "utils.dart",
+    error: uri,
+  );
+
+  if (uri.scheme != "otpauth") {
+    throw ArgumentError.value(
+      uri,
+      "uri",
+      "The uri is not a valid otpauth uri but a(n) [${uri.scheme}] uri instead.",
+    );
+  }
+
+  String type = uri.host;
+  if (equalsIgnoreCase(type, enumAsString(TokenTypes.HOTP)) ||
+      equalsIgnoreCase(type, enumAsString(TokenTypes.TOTP))) {
+    return parseOtpAuth(uri);
+  } else if (equalsIgnoreCase(type, enumAsString(TokenTypes.PIPUSH))) {
+    return parsePiAuth(uri);
+  }
+
+  throw ArgumentError.value(
+    uri,
+    "uri",
+    "The token type [$type] is not supported",
+  );
+}
+
+Map<String, dynamic> parsePiAuth(Uri uri) {
+  // otpauth://pipush/LABELTEXT?
+  // url=https://privacyidea.org/enroll/this/token
+  // &ttl=120
+  // &serial=PIPU0006EF87
+  // &projectid=test-d1231
+  // &appid=1:0123456789012:android:0123456789abcdef
+  // &apikey=AIzaSyBeFSjwJ8aEcHQaj4-isT-sLAX6lmSrvbb
+  // &projectnumber=850240559999
+  // &enrollment_credential=9311ee50678983c0f29d3d843f86e39405e2b427
+  // &apikeyios=AIzaSyBeFSjwJ8aEcHQaj4-isT-sLAX6lmSrvbb
+  // &appidios=1:0123456789012:ios:0123456789abcdef
+
+  Map<String, dynamic> uriMap = Map();
+
+  uriMap[URI_TYPE] = uri.host;
+  uriMap[URI_ISSUER] = uri.queryParameters['issuer'];
+
+  // If we do not support the version of this piauth url, we can stop here.
+  String pushVersionAsString = uri.queryParameters["v"];
+
+  if (pushVersionAsString == null) {
+    throw ArgumentError.value(uri, "uri",
+        "Parameter [v] is not an optional parameter and is missing.");
+  }
+
+  try {
+    int pushVersion = int.parse(pushVersionAsString);
+
+    print('VERSION: $pushVersion');
+
+    if (pushVersion > 1) {
+      throw ArgumentError.value(
+          uri,
+          "uri",
+          "The piauth version [$pushVersionAsString] "
+              "is not supported by this version of the app.");
+    }
+  } on FormatException {
+    throw ArgumentError.value(uri, "uri",
+        "[$pushVersionAsString] is not a valid value for parameter [v].");
+  }
+
+  uriMap[URI_LABEL] = _parseLabel(uri);
+
+  uriMap[URI_SERIAL] = uri.queryParameters["serial"];
+  ArgumentError.checkNotNull(uriMap[URI_SERIAL], "serial");
+
+  uriMap[URI_PROJECT_ID] = uri.queryParameters["projectid"];
+  ArgumentError.checkNotNull(uriMap[URI_PROJECT_ID], "projectid");
+
+  uriMap[URI_APP_ID] = uri.queryParameters["appid"];
+  ArgumentError.checkNotNull(uriMap[URI_APP_ID], "appid");
+
+  uriMap[URI_APP_ID_IOS] = uri.queryParameters["appidios"];
+  ArgumentError.checkNotNull(uriMap[URI_APP_ID_IOS], "appidios");
+
+  uriMap[URI_API_KEY] = uri.queryParameters["apikey"];
+  ArgumentError.checkNotNull(uriMap[URI_API_KEY], "apikey");
+
+  uriMap[URI_API_KEY_IOS] = uri.queryParameters["apikeyios"];
+  ArgumentError.checkNotNull(uriMap[URI_API_KEY_IOS], "apikeyios");
+
+  uriMap[URI_PROJECT_NUMBER] = uri.queryParameters["projectnumber"];
+  ArgumentError.checkNotNull(uriMap[URI_PROJECT_NUMBER], "projectnumber");
+
+  // TODO what happens if Uri.parse fails?
+  uriMap[URI_ROLLOUT_URL] = Uri.parse(uri.queryParameters["url"]);
+  ArgumentError.checkNotNull(uriMap[URI_ROLLOUT_URL], "url");
+
+  String ttlAsString = uri.queryParameters["ttl"] ?? "10";
+  try {
+    uriMap[URI_TTL] = int.parse(ttlAsString);
+  } on FormatException {
+    throw ArgumentError.value(
+        uri, "uri", "[$ttlAsString] is not a valid value for parameter [ttl].");
+  }
+
+  uriMap[URI_ENROLLMENT_CREDENTIAL] =
+      uri.queryParameters["enrollment_credential"];
+  ArgumentError.checkNotNull(
+      uriMap[URI_ENROLLMENT_CREDENTIAL], "enrollment_credential");
+
+  uriMap[URI_SSL_VERIFY] = (uri.queryParameters["sslverify"] ?? "1") == "1";
+
+  return uriMap;
+}
+
+/// This method parses otpauth uris according
+/// to https://github.com/google/google-authenticator/wiki/Key-Uri-Format.
+Map<String, dynamic> parseOtpAuth(Uri uri) {
+  // otpauth://TYPE/LABEL?PARAMETERS
+
+  Map<String, dynamic> uriMap = Map();
+
+  // parse.host -> Type totp or hotp
+  uriMap[URI_TYPE] = uri.host;
+  uriMap[URI_ISSUER] = uri.queryParameters['issuer'];
+
+// parse.path.substring(1) -> Label
+  print("Key: [..] | Value: [..]");
+  uri.queryParameters.forEach((key, value) {
+    print("  $key | $value");
+  });
+
+  uriMap[URI_LABEL] = _parseLabel(uri);
+
+  String algorithm = uri.queryParameters["algorithm"] ??
+      enumAsString(Algorithms.SHA1); // Optional parameter
+
+  if (!equalsIgnoreCase(algorithm, enumAsString(Algorithms.SHA1)) &&
+      !equalsIgnoreCase(algorithm, enumAsString(Algorithms.SHA256)) &&
+      !equalsIgnoreCase(algorithm, enumAsString(Algorithms.SHA512))) {
+    throw ArgumentError.value(
+      uri,
+      "uri",
+      "The algorithm [$algorithm] is not supported",
+    );
+  }
+
+  uriMap[URI_ALGORITHM] = algorithm;
+
+  // Parse digits.
+  String digitsAsString =
+      uri.queryParameters["digits"] ?? "6"; // Optional parameter
+
+  if (digitsAsString != "6" && digitsAsString != "8") {
+    throw ArgumentError.value(
+      uri,
+      "uri",
+      "[$digitsAsString] is not a valid number of digits",
+    );
+  }
+
+  int digits = int.parse(digitsAsString);
+
+  uriMap[URI_DIGITS] = digits;
+
+  // Parse secret.
+  String secretAsString = uri.queryParameters["secret"];
+
+  // This is a fix for omitted padding in base32 encoded secrets.
+  //
+  // According to https://github.com/google/google-authenticator/wiki/Key-Uri-Format,
+  // the padding can be omitted, but the libraries for base32 do not allow this.
+  if (secretAsString != null && secretAsString.length % 2 == 1) {
+    secretAsString += "=";
+  }
+
+  if (!isValidEncoding(secretAsString, Encodings.base32)) {
+    throw ArgumentError.value(
+      uri,
+      "uri",
+      "[${enumAsString(Encodings.base32)}] is not a valid encoding for [$secretAsString].",
+    );
+  }
+
+  Uint8List secret = decodeSecretToUint8(secretAsString, Encodings.base32);
+
+  uriMap[URI_SECRET] = secret;
+
+  if (uriMap[URI_TYPE] == "hotp") {
+    // Parse counter.
+    String counterAsString = uri.queryParameters["counter"];
+    try {
+      if (counterAsString == null) {
+        throw ArgumentError.value(
+          uri,
+          "uri",
+          "Value for parameter [counter] is not optional and is missing.",
+        );
+      }
+      uriMap[URI_COUNTER] = int.parse(counterAsString);
+    } on FormatException {
+      throw ArgumentError.value(
+        uri,
+        "uri",
+        "[$counterAsString] is not a valid value for uri parameter [counter].",
+      );
+    }
+  }
+
+  if (uriMap[URI_TYPE] == "totp") {
+    // Parse period.
+    String periodAsString = uri.queryParameters["period"] ?? "30";
+    if (periodAsString != "30" && periodAsString != "60") {
+      throw ArgumentError.value(
+        uri,
+        "uri",
+        "[$periodAsString] is not a valid value for the paramerter [period].",
+      );
+    }
+
+    uriMap[URI_PERIOD] = int.parse(periodAsString);
+  }
+
+  if (is2StepURI(uri)) {
+    // Parse for 2 step roll out.
+    String saltLengthAsString = uri.queryParameters["2step_salt"] ?? "10";
+    String outputLengthInByteAsString =
+        uri.queryParameters["2step_output"] ?? "20";
+    String iterationsAsString =
+        uri.queryParameters["2step_difficulty"] ?? "10000";
+
+    // Parse parameters
+    try {
+      uriMap[URI_SALT_LENGTH] = int.parse(saltLengthAsString);
+    } on FormatException {
+      throw ArgumentError.value(
+        uri,
+        "uri",
+        "[$saltLengthAsString] is not a valid value for parameter [2step_salt].",
+      );
+    }
+    try {
+      uriMap[URI_OUTPUT_LENGTH_IN_BYTES] =
+          int.parse(outputLengthInByteAsString);
+    } on FormatException {
+      throw ArgumentError.value(
+        uri,
+        "uri",
+        "[$outputLengthInByteAsString] is not a valid value for parameter [2step_output].",
+      );
+    }
+    try {
+      uriMap[URI_ITERATIONS] = int.parse(iterationsAsString);
+    } on FormatException {
+      throw ArgumentError.value(
+        uri,
+        "uri",
+        "[$iterationsAsString] is not a valid value for parameter [2step_difficulty].",
+      );
+    }
+  }
+
+  return uriMap;
+}
+
+String _parseLabel(Uri uri) {
+  return uri.path.substring(1);
+}
+
+bool is2StepURI(Uri uri) {
+  return uri.queryParameters["2step_salt"] != null ||
+      uri.queryParameters["2step_output"] != null ||
+      uri.queryParameters["2step_difficulty"] != null;
 }
