+++ test/utils/utils_test.dart
@@ -30,7 +30,6 @@ void main() {
   _testDecodeSecretToUint8();
   _testCalculateHotpValue();
   _testInsertCharAt();
-  _testParseOtpAuth();
 }
 
 void _testInsertCharAt() {
@@ -240,241 +239,6 @@ void _testDecodeSecretToUint8() {
     test("Test non base32 secret", () {
       expect(() => decodeSecretToUint8("AAAAAAöA", Encodings.base32),
           throwsA(TypeMatcher<FormatException>()));
-    });
-  });
-}
-
-void _testParseOtpAuth() {
-  group("parse otpauth uris: ", () {
-    group("HOTP and TOTP", () {
-      test("Test with wrong uri schema", () {
-        expect(
-            () => parseQRCodeToMap("http://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=SHA1&digits=6&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test with unknown type", () {
-        expect(
-            () => parseQRCodeToMap("otpauth://asdf/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=SHA1&digits=6&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test with missing type", () {
-        expect(
-            () => parseQRCodeToMap("otpauth:///ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=SHA1&digits=6&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test missing algorithm", () {
-        Map<String, dynamic> map =
-            parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&digits=6&period=30");
-        expect(map[URI_ALGORITHM], "SHA1"); // This is the default value
-      });
-
-      test("Test unknown algorithm", () {
-        expect(
-            () => parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=BubbleSort&digits=6&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test missing digits", () {
-        Map<String, dynamic> map =
-            parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&period=30");
-        expect(map[URI_DIGITS], 6); // This is the default value
-      });
-
-      // At least the library used to calculate otp values does not support other number of digits.
-      test("Test invalid number of digits", () {
-        expect(
-            () => parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=SHA1&digits=66&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test invalid characters for digits", () {
-        expect(
-            () => parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=SHA1&digits=aA&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test missing secret", () {
-        expect(
-            () => parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "issuer=ACME%20Co&algorithm=SHA1&digits=6&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test invalid secret", () {
-        expect(
-            () => parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=ÖÖ&issuer=ACME%20Co&algorithm=SHA1&digits=6"
-                "&period=30"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      // TOTP specific
-      test("Test missing period", () {
-        Map<String, dynamic> map =
-            parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=SHA1&digits=6");
-        expect(map[URI_PERIOD], 30);
-      });
-
-      test("Test invalid characters for period", () {
-        expect(
-            () => parseQRCodeToMap("otpauth://totp/ACME%20Co:john@example.com?"
-                "secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ&issuer=ACME%20Co"
-                "&algorithm=SHA1&digits=6&period=aa"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test valid totp uri", () {
-        Map<String, dynamic> map = parseQRCodeToMap(
-            "otpauth://totp/Kitchen?secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ"
-            "&issuer=ACME%20Co&algorithm=SHA512&digits=8&period=60");
-        expect(map[URI_LABEL], "Kitchen");
-        expect(map[URI_ALGORITHM], "SHA512");
-        expect(map[URI_DIGITS], 8);
-        expect(
-            map[URI_SECRET],
-            decodeSecretToUint8(
-                "HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ", Encodings.base32));
-        expect(map[URI_PERIOD], 60);
-      });
-
-      // HOTP specific
-      test("Test with missing counter", () {
-        expect(
-            () => parseQRCodeToMap(
-                "otpauth://hotp/Kitchen?secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ"
-                "&issuer=ACME%20Co&algorithm=SHA256&digits=8"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test with invalid counter", () {
-        expect(
-            () => parseQRCodeToMap(
-                "otpauth://hotp/Kitchen?secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ"
-                "&issuer=ACME%20Co&algorithm=SHA256&digits=8&counter=aa"),
-            throwsA(TypeMatcher<ArgumentError>()));
-      });
-
-      test("Test valid hotp uri", () {
-        Map<String, dynamic> map = parseQRCodeToMap(
-            "otpauth://hotp/Kitchen?secret=HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ"
-            "&issuer=ACME%20Co&algorithm=SHA256&digits=8&counter=5");
-        expect(map[URI_LABEL], "Kitchen");
-        expect(map[URI_ALGORITHM], "SHA256");
-        expect(map[URI_DIGITS], 8);
-        expect(
-            map[URI_SECRET],
-            decodeSecretToUint8(
-                "HXDMVJECJJWSRB3HWIZR4IFUGFTMXBOZ", Encodings.base32));
-        expect(map[URI_COUNTER], 5);
-      });
-    });
-
-    group("2 Step Rollout", () {
-      test("is2StepURI", () {
-        expect(
-            is2StepURI(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE&counter=1&digits=6&issuer=privacyIDEA&2step_salt=8&2step_output=20")),
-            true);
-        expect(
-            is2StepURI(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE&counter=1&digits=6&issuer=privacyIDEA&2step_salt=8&2step_difficulty=10000")),
-            true);
-        expect(
-            is2StepURI(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE&counter=1&digits=6&issuer=privacyIDEA&2step_output=20&2step_difficulty=10000")),
-            true);
-        expect(
-            is2StepURI(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE&counter=1&digits=6&issuer=privacyIDEA&2step_salt=8")),
-            true);
-
-        expect(
-            is2StepURI(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE&counter=1&digits=6&issuer=privacyIDEA")),
-            false);
-      });
-
-      test("parse complete uri", () {
-        Map<String, dynamic> uriMap = parseOtpAuth(Uri.parse(
-            "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE"
-            "&counter=1&digits=6&issuer=privacyIDEA&2step_salt=54"
-            "&2step_output=42&2step_difficulty=12345"));
-
-        expect(uriMap[URI_SALT_LENGTH], 54);
-        expect(uriMap[URI_OUTPUT_LENGTH_IN_BYTES], 42);
-        expect(uriMap[URI_ITERATIONS], 12345);
-      });
-
-      test("parse with default values", () {
-        expect(
-            parseOtpAuth(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE"
-                "&counter=1&digits=6&issuer=privacyIDEA&2step_output=42"
-                "&2step_difficulty=12345"))[URI_SALT_LENGTH],
-            10);
-        expect(
-            parseOtpAuth(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE"
-                "&counter=1&digits=6&issuer=privacyIDEA&2step_salt=54"
-                "&2step_difficulty=12345"))[URI_OUTPUT_LENGTH_IN_BYTES],
-            20);
-        expect(
-            parseOtpAuth(Uri.parse(
-                "otpauth://hotp/OATH0001F662?secret=HDOMWJ5GEQQA6RR34RAP55QBVCX3E2RE"
-                "&counter=1&digits=6&issuer=privacyIDEA&2step_salt=54"
-                "&2step_output=42"))[URI_ITERATIONS],
-            10000);
-      });
-    });
-
-    group("Push Token", () {
-      test("parse complete uri", () {
-        // TODO Do these work on iOS?
-
-        Map<String, dynamic> uriMap = parsePiAuth(Uri.parse(
-            "otpauth://pipush/PIPU0001353C?url=https%3A//192.168.178.32/ttype/"
-            "push&ttl=2&issuer=privacyIDEA"
-            "&enrollment_credential=69ebdbd37a70dcd2dfd4d166cc5325158dc7befe"
-            "&projectnumber=645909237054&projectid=lkfhgdf"
-            "&appid=1%3A645909237054%3Aandroid%3A812605f9a33242a9"
-            "&apikey=AIzaSyAGyFpiB9pUILcXKpfagENJCPqvvUWdiRk"
-            "&appidios=AIzaSyAGyFpiB9pUILcXKpfagENJCPqvvUWdiRk"
-            "&apikeyios=AIzaSyAGyFpiB9pUILcXKpfagENJCPqvvUWdiRk&v=1"
-            "&serial=PIPU0001353C&sslverify=0"));
-
-        expect(uriMap[URI_LABEL], "PIPU0001353C");
-        expect(uriMap[URI_SERIAL], "PIPU0001353C");
-        expect(uriMap[URI_TTL], 2);
-        expect(uriMap[URI_ISSUER], "privacyIDEA");
-        expect(uriMap[URI_ENROLLMENT_CREDENTIAL],
-            "69ebdbd37a70dcd2dfd4d166cc5325158dc7befe");
-        expect(uriMap[URI_PROJECT_NUMBER], "645909237054");
-        expect(uriMap[URI_PROJECT_ID], "lkfhgdf");
-        expect(uriMap[URI_APP_ID], "1:645909237054:android:812605f9a33242a9");
-        expect(uriMap[URI_API_KEY], "AIzaSyAGyFpiB9pUILcXKpfagENJCPqvvUWdiRk");
-        expect(uriMap[URI_SSL_VERIFY], false);
-      });
     });
   });
 }
